# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sinovera Transit Global API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List warehouses
      responses:
        "200":
          description: Warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"

    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create warehouse
      security: [{ clerkAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Created warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /warehouses/{id}:
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update warehouse
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"

    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      summary: Delete warehouse
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Carriers ───────────────────────────────────────────────────────────────
  /carriers:
    get:
      operationId: listCarriers
      tags: [carriers]
      summary: List carriers
      responses:
        "200":
          description: Carriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Carrier"

    post:
      operationId: createCarrier
      tags: [carriers]
      summary: Create carrier
      security: [{ clerkAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierInput"
      responses:
        "201":
          description: Created carrier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
  /carriers/{id}:
    patch:
      operationId: updateCarrier
      tags: [carriers]
      summary: Update carrier
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"

    delete:
      operationId: deleteCarrier
      tags: [carriers]
      summary: Delete carrier
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── CMS ────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Warehouse:
      type: object
      required: [id, name, code, address, city, country, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    WarehouseInput:
      type: object
      required: [name, code, address, city, country]
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
    WarehouseUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        isActive:
          type: boolean

    # ── Carrier ───────────────────────────────────────────────────────────────
    Carrier:
      type: object
      required: [id, name, code, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        trackingUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CarrierInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        trackingUrl:
          type: string
    CarrierUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        trackingUrl:
          type: string
        isActive:
          type: boolean

    # ── CMS ───────────────────────────────────────────────────────────────────
...